Separate api and domain models to own project modules
