+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/request/ActivationCodeRequest.java
@@ -15,16 +15,19 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.request;
+package com.wultra.app.enrollmentserver.api.model.request;
 
 import lombok.Data;
 
 /**
- * Request class used when initializing presence check.
+ * Request with attributes of a new activation code.
  *
- * @author Roman Strobl, roman.strobl@wultra.com
+ * @author Petr Dvorak, petr@wultra.com
  */
 @Data
-public class PresenceCheckInitRequest {
+public class ActivationCodeRequest {
+
+    public String applicationId;
+    public String otp;
 
 }
