+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/request/IdentityVerificationInitRequest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.request;
+package com.wultra.app.enrollmentserver.api.model.request;
 
 import lombok.Data;
