+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/request/PresenceCheckInitRequest.java
@@ -15,18 +15,16 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.request;
+package com.wultra.app.enrollmentserver.api.model.request;
 
 import lombok.Data;
 
 /**
- * Request class used for cleanup related to onboarding process.
+ * Request class used when initializing presence check.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class OnboardingCleanupRequest {
-
-    private String processId;
+public class PresenceCheckInitRequest {
 
 }
