+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/request/PushRegisterRequest.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Enrollment Server
- * Copyright (C) 2021 Wultra s.r.o.
+ * Copyright (C) 2020 Wultra s.r.o.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -15,27 +15,22 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.request;
+
+package com.wultra.app.enrollmentserver.api.model.request;
 
 import lombok.Data;
 
-import java.util.List;
-
 /**
- * Request class used when checking identity document verification status.
+ * Class representing a device registration request. The supported platform
+ * values are 'ios' and 'android'. The push token is the value received from
+ * APNS or FCM services without any modification.
  *
- * @author Roman Strobl, roman.strobl@wultra.com
+ * @author Petr Dvorak, petr@wultra.com
  */
 @Data
-public class DocumentStatusRequest {
+public class PushRegisterRequest {
 
-    private List<DocumentFilter> filter;
-
-    @Data
-    public static class DocumentFilter {
-
-        private String documentId;
-
-    }
+    private String platform;
+    private String token;
 
 }
