+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/response/DocumentStatusResponse.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Enrollment Server
- * Copyright (C) 2020 Wultra s.r.o.
+ * Copyright (C) 2021 Wultra s.r.o.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -15,22 +15,33 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+package com.wultra.app.enrollmentserver.api.model.response;
 
-package com.wultra.app.enrollmentserver.model.request;
-
+import com.wultra.app.enrollmentserver.model.enumeration.DocumentStatus;
+import com.wultra.app.enrollmentserver.model.enumeration.IdentityVerificationStatus;
 import lombok.Data;
 
+import java.util.List;
+
 /**
- * Class representing a device registration request. The supported platform
- * values are 'ios' and 'android'. The push token is the value received from
- * APNS or FCM services without any modification.
+ * Response class used when checking identity document verification status.
  *
- * @author Petr Dvorak, petr@wultra.com
+ * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class PushRegisterRequest {
+public class DocumentStatusResponse {
 
-    private String platform;
-    private String token;
+    private IdentityVerificationStatus status;
+    private List<DocumentMetadata> documents;
+
+    @Data
+    public static class DocumentMetadata {
+
+        private String filename;
+        private String id;
+        private DocumentStatus status;
+        private List<String> errors;
+
+    }
 
 }
