+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/response/DocumentSubmitResponse.java
@@ -15,20 +15,35 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.response;
+package com.wultra.app.enrollmentserver.api.model.response;
 
+import com.wultra.app.enrollmentserver.model.enumeration.CardSide;
+import com.wultra.app.enrollmentserver.model.enumeration.DocumentStatus;
+import com.wultra.app.enrollmentserver.model.enumeration.DocumentType;
 import lombok.Data;
 
+import java.util.List;
+
 /**
- * Response with a new activation code.
+ * Response class used when submitting documents for identity verification.
  *
- * @author Petr Dvorak, petr@wultra.com
+ * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class ActivationCodeResponse {
+public class DocumentSubmitResponse {
 
-    private String activationId;
-    private String activationCode;
-    private String activationSignature;
+    private List<DocumentMetadata> documents;
+
+    @Data
+    public static class DocumentMetadata {
+
+        private String filename;
+        private String id;
+        private DocumentType type;
+        private CardSide side;
+        private DocumentStatus status;
+        private List<String> errors;
+
+    }
 
 }
