+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/response/DocumentUploadResponse.java
@@ -15,18 +15,19 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.request;
+package com.wultra.app.enrollmentserver.api.model.response;
 
 import lombok.Data;
 
 /**
- * Request class used when checking onboarding process status.
+ * Response class used when uploading documents for identity verification.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class OnboardingStatusRequest {
+public class DocumentUploadResponse {
 
-    private String processId;
+    private String filename;
+    private String id;
 
 }
