+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/response/IdentityVerificationStatusResponse.java
@@ -15,22 +15,21 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.response;
+package com.wultra.app.enrollmentserver.api.model.response;
 
-import com.wultra.app.enrollmentserver.model.enumeration.OnboardingStatus;
+import com.wultra.app.enrollmentserver.model.enumeration.IdentityVerificationPhase;
+import com.wultra.app.enrollmentserver.model.enumeration.IdentityVerificationStatus;
 import lombok.Data;
 
-import java.util.Map;
-
 /**
- * Response class used when starting the onboarding process.
+ * Response class used when checking identity verification status.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class OnboardingStartResponse {
+public class IdentityVerificationStatusResponse {
 
-    private String processId;
-    private OnboardingStatus onboardingStatus;
+    private IdentityVerificationStatus identityVerificationStatus;
+    private IdentityVerificationPhase identityVerificationPhase;
 
 }
