+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/response/OnboardingStartResponse.java
@@ -15,18 +15,20 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.response;
+package com.wultra.app.enrollmentserver.api.model.response;
 
 import com.wultra.app.enrollmentserver.model.enumeration.OnboardingStatus;
 import lombok.Data;
 
+import java.util.Map;
+
 /**
- * Response class used when checking onboarding process status.
+ * Response class used when starting the onboarding process.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class OnboardingStatusResponse {
+public class OnboardingStartResponse {
 
     private String processId;
     private OnboardingStatus onboardingStatus;
