+++ enrollment-server-api-model/src/main/java/com/wultra/app/enrollmentserver/api/model/response/PresenceCheckInitResponse.java
@@ -15,22 +15,20 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.app.enrollmentserver.model.response;
+package com.wultra.app.enrollmentserver.api.model.response;
 
-import com.wultra.app.enrollmentserver.model.enumeration.OnboardingStatus;
 import lombok.Data;
 
+import java.util.Map;
+
 /**
- * Response class used when verifying an OTP code.
+ * Response class used when initializing presence check.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
-public class OtpVerifyResponse {
+public class PresenceCheckInitResponse {
 
-    private String processId;
-    private OnboardingStatus onboardingStatus;
-    private boolean verified;
-    private Integer remainingAttempts;
+    private Map<String, Object> sessionAttributes;
 
 }
