+++ enrollment-server-domain-model/src/main/java/com/wultra/app/enrollmentserver/model/integration/DocumentVerificationResult.java
