+++ enrollment-server/src/main/java/com/wultra/app/enrollmentserver/controller/api/ActivationCodeController.java
@@ -20,8 +20,8 @@ package com.wultra.app.enrollmentserver.controller.api;
 import com.wultra.app.enrollmentserver.errorhandling.ActivationCodeException;
 import com.wultra.app.enrollmentserver.errorhandling.InvalidRequestObjectException;
 import com.wultra.app.enrollmentserver.impl.service.ActivationCodeService;
-import com.wultra.app.enrollmentserver.model.request.ActivationCodeRequest;
-import com.wultra.app.enrollmentserver.model.response.ActivationCodeResponse;
+import com.wultra.app.enrollmentserver.api.model.request.ActivationCodeRequest;
+import com.wultra.app.enrollmentserver.api.model.response.ActivationCodeResponse;
 import io.getlime.core.rest.model.base.request.ObjectRequest;
 import io.getlime.core.rest.model.base.response.ObjectResponse;
 import io.getlime.security.powerauth.crypto.lib.encryptor.ecies.model.EciesScope;
