+++ enrollment-server/src/main/java/com/wultra/app/enrollmentserver/controller/api/OnboardingController.java
@@ -20,18 +20,17 @@ package com.wultra.app.enrollmentserver.controller.api;
 import com.wultra.app.enrollmentserver.errorhandling.OnboardingOtpDeliveryException;
 import com.wultra.app.enrollmentserver.errorhandling.OnboardingProcessException;
 import com.wultra.app.enrollmentserver.impl.service.OnboardingService;
-import com.wultra.app.enrollmentserver.model.request.OnboardingCleanupRequest;
-import com.wultra.app.enrollmentserver.model.request.OnboardingStartRequest;
-import com.wultra.app.enrollmentserver.model.request.OnboardingStatusRequest;
-import com.wultra.app.enrollmentserver.model.request.OtpResendRequest;
-import com.wultra.app.enrollmentserver.model.response.OnboardingStartResponse;
-import com.wultra.app.enrollmentserver.model.response.OnboardingStatusResponse;
+import com.wultra.app.enrollmentserver.api.model.request.OnboardingCleanupRequest;
+import com.wultra.app.enrollmentserver.api.model.request.OnboardingStartRequest;
+import com.wultra.app.enrollmentserver.api.model.request.OnboardingStatusRequest;
+import com.wultra.app.enrollmentserver.api.model.request.OtpResendRequest;
+import com.wultra.app.enrollmentserver.api.model.response.OnboardingStartResponse;
+import com.wultra.app.enrollmentserver.api.model.response.OnboardingStatusResponse;
 import io.getlime.core.rest.model.base.request.ObjectRequest;
 import io.getlime.core.rest.model.base.response.ObjectResponse;
 import io.getlime.core.rest.model.base.response.Response;
 import io.getlime.security.powerauth.crypto.lib.encryptor.ecies.model.EciesScope;
 import io.getlime.security.powerauth.rest.api.spring.encryption.EciesEncryptionContext;
-import io.getlime.security.powerauth.rest.api.spring.exception.PowerAuthAuthenticationException;
 import io.getlime.security.powerauth.rest.api.spring.annotation.EncryptedRequestBody;
 import io.getlime.security.powerauth.rest.api.spring.annotation.PowerAuthEncryption;
 import io.getlime.security.powerauth.rest.api.spring.exception.PowerAuthEncryptionException;
