+++ enrollment-server/src/main/java/com/wultra/app/enrollmentserver/impl/service/ActivationCodeService.java
@@ -20,8 +20,8 @@ package com.wultra.app.enrollmentserver.impl.service;
 import com.wultra.app.enrollmentserver.errorhandling.ActivationCodeException;
 import com.wultra.app.enrollmentserver.errorhandling.InvalidRequestObjectException;
 import com.wultra.app.enrollmentserver.impl.service.converter.ActivationCodeConverter;
-import com.wultra.app.enrollmentserver.model.request.ActivationCodeRequest;
-import com.wultra.app.enrollmentserver.model.response.ActivationCodeResponse;
+import com.wultra.app.enrollmentserver.api.model.request.ActivationCodeRequest;
+import com.wultra.app.enrollmentserver.api.model.response.ActivationCodeResponse;
 import com.wultra.app.enrollmentserver.model.validator.ActivationCodeRequestValidator;
 import com.wultra.security.powerauth.client.PowerAuthClient;
 import com.wultra.security.powerauth.client.model.error.PowerAuthClientException;
